bug - even after deleting it. The assignment still is there

+++ routes/assignment.route.js
@@ -182,11 +182,21 @@ assignmentRouter.delete(
   authorize(["instructor", "admin"]),
   async (req, res) => {
     try {
+      const isAssignmentExist = await Assignment.findOne({
+        where: {
+          id: req.params.id,
+        },
+      });
+
+      if (!isAssignmentExist)
+        return res.status(404).json({ message: "Assignment does not exist." });
+
       await Assignment.destroy({
         where: {
           id: req.params.id,
         },
       });
+
       res.json({ message: "Assignment Deleted." });
     } catch (err) {
       console.error("Error fetching assignments:", err);
@@ -205,6 +215,15 @@ assignmentRouter.patch(
 
       if (updatedAssignment.id)
         return res.json({ message: "Exclude the assignment id!" });
+
+      const isAssignmentExist = await Assignment.findOne({
+        where: {
+          id: req.params.id,
+        },
+      });
+
+      if (!isAssignmentExist)
+        return res.status(404).json({ message: "Assignment does not exist." });
 
       await Assignment.update(updatedAssignment, {
         where: {
